requirements.txt
solana~=0.28.0
base58~=2.1.1
requests==2.28.1
solders~=0.10.0
colorama~=0.4.4
python-dotenv~=0.21.0
discord
py-cord